コード一貫性を向上させるリファクタリング
CLAUDE.mdルール準拠:
- interface → type 変換 (color-mode.tsx, tooltip.tsx)
- Function Declaration → Arrow Function 変換
- Promise.then → async/await 変換 (reportWebVitals.ts)

エラーハンドリング統一:
- console.error 削除、toaster.create に統一
- pages/Invite の責務分離 (useMutation を features に移動)
- SendTab, UserRegister のエラーハンドリング統一

コンポーネント改善:
- LoadingState 共通コンポーネント作成
- TopPage をセクションコンポーネントに分割 (643行 → 33行)
- ShiftTemplateList の JSX ネスト削減と命名改善

アクセシビリティ改善:
- ShiftTemplateForm の曜日選択を Badge → Button に変更
- MyPage のお知らせカードを Box → Button に変更
