+++ src/components/config/color-mode.tsx
@@ -5,21 +5,21 @@ import { ThemeProvider, useTheme } from "next-themes";
 import * as React from "react";
 import { LuMoon, LuSun } from "react-icons/lu";
 
-export interface ColorModeProviderProps extends ThemeProviderProps {}
+export type ColorModeProviderProps = ThemeProviderProps;
 
-export function ColorModeProvider(props: ColorModeProviderProps) {
+export const ColorModeProvider = (props: ColorModeProviderProps) => {
   return <ThemeProvider attribute="class" disableTransitionOnChange {...props} forcedTheme="light" />;
-}
+};
 
 export type ColorMode = "light" | "dark";
 
-export interface UseColorModeReturn {
+export type UseColorModeReturn = {
   colorMode: ColorMode;
   setColorMode: (colorMode: ColorMode) => void;
   toggleColorMode: () => void;
-}
+};
 
-export function useColorMode(): UseColorModeReturn {
+export const useColorMode = (): UseColorModeReturn => {
   const { resolvedTheme, setTheme, forcedTheme } = useTheme();
   const colorMode = forcedTheme || resolvedTheme;
   const toggleColorMode = () => {
@@ -30,47 +30,45 @@ export function useColorMode(): UseColorModeReturn {
     setColorMode: setTheme,
     toggleColorMode,
   };
-}
+};
 
-export function useColorModeValue<T>(light: T, dark: T) {
+export const useColorModeValue = <T,>(light: T, dark: T) => {
   const { colorMode } = useColorMode();
   return colorMode === "dark" ? dark : light;
-}
+};
 
-export function ColorModeIcon() {
+export const ColorModeIcon = () => {
   const { colorMode } = useColorMode();
   return colorMode === "dark" ? <LuMoon /> : <LuSun />;
-}
+};
 
-interface ColorModeButtonProps extends Omit<IconButtonProps, "aria-label"> {}
+type ColorModeButtonProps = Omit<IconButtonProps, "aria-label">;
 
-export const ColorModeButton = React.forwardRef<HTMLButtonElement, ColorModeButtonProps>(
-  function ColorModeButton(props, ref) {
-    const { toggleColorMode } = useColorMode();
-    return (
-      <ClientOnly fallback={<Skeleton boxSize="9" />}>
-        <IconButton
-          onClick={toggleColorMode}
-          variant="ghost"
-          aria-label="Toggle color mode"
-          size="sm"
-          ref={ref}
-          {...props}
-          css={{
-            _icon: {
-              width: "5",
-              height: "5",
-            },
-          }}
-        >
-          <ColorModeIcon />
-        </IconButton>
-      </ClientOnly>
-    );
-  },
-);
+export const ColorModeButton = React.forwardRef<HTMLButtonElement, ColorModeButtonProps>((props, ref) => {
+  const { toggleColorMode } = useColorMode();
+  return (
+    <ClientOnly fallback={<Skeleton boxSize="9" />}>
+      <IconButton
+        onClick={toggleColorMode}
+        variant="ghost"
+        aria-label="Toggle color mode"
+        size="sm"
+        ref={ref}
+        {...props}
+        css={{
+          _icon: {
+            width: "5",
+            height: "5",
+          },
+        }}
+      >
+        <ColorModeIcon />
+      </IconButton>
+    </ClientOnly>
+  );
+});
 
-export const LightMode = React.forwardRef<HTMLSpanElement, SpanProps>(function LightMode(props, ref) {
+export const LightMode = React.forwardRef<HTMLSpanElement, SpanProps>((props, ref) => {
   return (
     <Span
       color="fg"
@@ -84,7 +82,7 @@ export const LightMode = React.forwardRef<HTMLSpanElement, SpanProps>(function L
   );
 });
 
-export const DarkMode = React.forwardRef<HTMLSpanElement, SpanProps>(function DarkMode(props, ref) {
+export const DarkMode = React.forwardRef<HTMLSpanElement, SpanProps>((props, ref) => {
   return (
     <Span
       color="fg"
