+++ src/components/features/User/Join/index.stories.tsx
@@ -23,9 +23,7 @@ type Story = StoryObj<typeof meta>;
 export const Basic: Story = {
   args: {
     invitation: mockInvitation,
-    error: null,
-    isAccepting: false,
-    onAccept: () => {},
+    token: "mock-token-123",
   },
 };
 
@@ -75,34 +73,13 @@ export const RequireLoginStory: StoryObj<typeof RequireLogin> = {
   render: () => <RequireLogin />,
 };
 
-// ログイン済み - エラーあり
-export const LoggedInWithError: StoryObj<typeof LoggedIn> = {
-  render: () => (
-    <LoggedIn
-      invitation={mockInvitation}
-      error="参加処理に失敗しました。もう一度お試しください。"
-      isAccepting={false}
-      onAccept={() => {}}
-    />
-  ),
-};
-
-// ログイン済み - 処理中
-export const LoggedInAccepting: StoryObj<typeof LoggedIn> = {
-  render: () => <LoggedIn invitation={mockInvitation} error={null} isAccepting={true} onAccept={() => {}} />,
-};
-
 // マネージャー役割
-export const ManagerRole: StoryObj<typeof LoggedIn> = {
-  render: () => (
-    <LoggedIn
-      invitation={{
-        ...mockInvitation,
-        role: "manager",
-      }}
-      error={null}
-      isAccepting={false}
-      onAccept={() => {}}
-    />
-  ),
+export const ManagerRole: Story = {
+  args: {
+    invitation: {
+      ...mockInvitation,
+      role: "manager",
+    },
+    token: "mock-token-manager",
+  },
 };
