+++ src/components/pages/Invite/index.tsx
@@ -1,60 +1,15 @@
 import { useAuth } from "@clerk/clerk-react";
-import { useMutation, useQuery } from "convex/react";
-import { useState } from "react";
+import { useQuery } from "convex/react";
 import { api } from "@/convex/_generated/api";
-import {
-  Accepted,
-  AlreadyAccepted,
-  ErrorView,
-  Loading,
-  LoggedIn,
-  RequireLogin,
-} from "@/src/components/features/User/Join";
+import { AlreadyAccepted, ErrorView, Loading, LoggedIn, RequireLogin } from "@/src/components/features/User/Join";
 
 type Props = {
   token: string;
 };
 
 export const InvitePage = ({ token }: Props) => {
-  const { isSignedIn, userId } = useAuth();
-  const [isAccepting, setIsAccepting] = useState(false);
-  const [isAccepted, setIsAccepted] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const [acceptedShopId, setAcceptedShopId] = useState<string | null>(null);
-  const [acceptedShopName, setAcceptedShopName] = useState<string | null>(null);
-
+  const { isSignedIn } = useAuth();
   const invitation = useQuery(api.invite.getInvitationByToken, token ? { token } : "skip");
-  const acceptInvitation = useMutation(api.invite.acceptInvitation);
-
-  // 承認処理
-  const handleAccept = async () => {
-    if (!userId) return;
-
-    setIsAccepting(true);
-    setError(null);
-
-    try {
-      const result = await acceptInvitation({
-        token,
-        authId: userId,
-      });
-
-      if (result.success) {
-        setIsAccepted(true);
-        setAcceptedShopId(result.data.shopId);
-        setAcceptedShopName(result.data.shopName);
-      }
-    } catch (e) {
-      setError(e instanceof Error ? e.message : "参加処理に失敗しました");
-    } finally {
-      setIsAccepting(false);
-    }
-  };
-
-  // 承認完了後の表示（最優先でチェック）
-  if (isAccepted && acceptedShopId && acceptedShopName) {
-    return <Accepted shopId={acceptedShopId} shopName={acceptedShopName} />;
-  }
 
   // トークンがない場合
   if (!token) {
@@ -98,6 +53,6 @@ export const InvitePage = ({ token }: Props) => {
     return <RequireLogin />;
   }
 
-  // ログイン済み - 正常系
-  return <LoggedIn invitation={invitation} error={error} isAccepting={isAccepting} onAccept={handleAccept} />;
+  // ログイン済み - 正常系（useMutationはLoggedIn内で管理）
+  return <LoggedIn invitation={invitation} token={token} />;
 };
